+++ streams/shared/src/main/scala/zio/stream/encoding1/ZTransducer1.scala
@@ -1,4 +1,4 @@
-package zio.stream.experiment1
+package zio.stream.encoding1
 
 import zio._
