+++ streams/shared/src/main/scala/zio/stream/encoding2/ZStream2.scala
@@ -1,4 +1,4 @@
-package zio.stream.experiment2
+package zio.stream.encoding2
 
 import zio._
